Move more methods into avro impl
